% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_tracklibs.R
\name{join_ridefiles}
\alias{join_ridefiles}
\title{Join ride segments which come from separate files}
\usage{
join_ridefiles(joinlist, summaries, tracks)
}
\arguments{
\item{joinlist}{list of vectors (order matters within vectors) of .fit
or .gpx filenames, each vector a part of a track to be joined in order}

\item{summaries}{tibble of ride summaries created from gpx and/or fit files}

\item{tracks}{tibble of ride tracks created from gpx and/or fit files}
}
\value{
a list of two tibbles containing the summaries and tracks with the
   specified sets of tracks merged
}
\description{
\code{join_ridefiles} Join two or more ride segments which come from
  separate files, since binary .fit files are not amenable to easy
  editing.  The merged data is in the updated entries in the
  (summary,track) data pair.  Tee joined .fit files cannot be
  exported.
}
\seealso{
\code{\link{read_ride}}
}
