% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_visuals.R
\name{map_rides}
\alias{map_rides}
\title{plot ride track data on a map}
\usage{
map_rides(geodf, maptitle, rgl = FALSE, ...)
}
\arguments{
\item{geodf}{data frame or tibble containing at least: position_lat.dd,
position_lon.dd,(or lat,lon)(both numeric,decimal degrees),
startbutton.date(int),startbutton.time(int),segment(numeric),
and speed.m.s for speed coloring}

\item{maptitle}{string containing title}

\item{rgl}{use rgl openGL to draw 3D track in viewer}

\item{...}{arguments for 3d rendering calls}
}
\description{
\code{map_rides} Plot lat/long data on maps, optionally using colored
  dots to indicate speed
}
\details{
Create a map displaying the tracks specified.  Tracks may be drawn all in one
  specified color, with each separate track drawn in a different color from
  a specified palette, or as a series of closely spaced dots (which may
  be prodded into appearing as a line through judicious choice of point
  size and alpha) with the color varying based on speed and the chosen
  palette.
}
