% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapWindow.R
\name{mapWindow}
\alias{mapWindow}
\title{plot ride track data on a map}
\usage{
mapWindow(geodf, definedmaps, usemap, margin.factor = 0.05, ...)
}
\arguments{
\item{geodf}{data frame or tibble containing at least: position_lat.dd,
position_lon.dd}

\item{definedmaps}{list of name,bounds pairs of maps}

\item{usemap}{name of map in list of defined maps}

\item{margin.factor}{percentage to expand map area beyond limits of
the track in each direction, when autodetermining region to plot}

\item{...}{other arguments}
}
\value{
c(min lon, max lon, min lat, max lat)
}
\description{
\code{mapWindow} Return a 4-tuple box from tracks and maps
}
\details{
Create a 4-tuple rectangle of lons and lats that is a specified map
  in a list of maps or is a minimal enclosure of the track pts
}
