% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_visuals.R
\name{plot_elev_profile_plus}
\alias{plot_elev_profile_plus}
\title{Plot ride track data on a map}
\usage{
plot_elev_profile_plus(track, summary, savefn, title = "Ride",
  palette = "magma", vertical.multiplier = NA, ppm = NA,
  show.stops = TRUE, show.time = TRUE, show.summary = TRUE,
  hr.low = 125, hr.high = 170, hr.color.low = 10, hr.color.high = 33,
  cad.low = 65, cad.target = 82, cad.color.low = 4, cad.color.mid = 8,
  cad.color.high = 15, hr.smooth.bw.meters = 12, hr.smooth.nn = 8,
  cad.smooth.bw.meters = 30, cad.smooth.nn = 12, min.numpoints = 1000)
}
\arguments{
\item{track}{data frame or tibble containing track details including
location and elevation data, as well as heartrate and cadence}

\item{summary}{data frame or tibble containing track summary from the
data frames generated by \code{read_ride}}

\item{savefn}{name of output file, extension is either .pdf or .png}

\item{title}{title printed over top center of map}

\item{palette}{palette to use in graphicv(\code{"magma"} or
\code{"inferno"} or \code{"plasma"})}

\item{vertical.multiplier}{default vertical exaggeration factor.  Default
ranges from 25 to 60 depending on the length being plotted}

\item{ppm}{override calculated default number of points per mile}

\item{show.stops}{draw row with short and long stops}

\item{show.time}{draw row with 15 minute and hour marks}

\item{show.summary}{display summary results if supplied}

\item{hr.low}{heartrates below this are same color}

\item{hr.high}{heartrates above this are same color}

\item{hr.color.low}{set color for hr.low and lower}

\item{hr.color.high}{set color for hr.highand higher
colors are from same palette as speeds, number is the speed corresponding
to the desired limit on the range of heartrates}

\item{cad.low}{low cadence limit}

\item{cad.target}{target cadence range minimum}

\item{cad.color.low}{set color for cadence at cad.low or below}

\item{cad.color.mid}{set color for cadence above low but below target}

\item{cad.color.high}{set color for cadence above target}

\item{hr.smooth.bw.meters}{bandwidth (in meters) for smoothing kernel
for heartrate data}

\item{hr.smooth.nn}{number of points (on each side) to use in}

\item{cad.smooth.bw.meters}{bandwidth (in meters) for smoothing kernel
for cadence data}

\item{cad.smooth.nn}{number of points (on each side) to use in
smoothing kernel for cadence data}

\item{min.numpoints}{pad out the plot on the right if too short}
}
\value{
a ggplot object
}
\description{
\code{plot_elev_profile_plus} creates a plot which uses the elevation profile
  and colored bands to compactly display the relationship between terrain,
  speed, cadence and heartrate on a given ride
}
\details{
Generates a plot which compactly displays the elevation profile and its
   relationship with speed, cadence and heartrate.  Also displays clock
   information and shows location and approximate length of stops.
}
