% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_visuals.R
\name{plot_profile}
\alias{plot_profile}
\title{plot heartrate, cadence, speed and more using color-driven
  data summaries to convey performance patterns}
\usage{
plot_profile(track, summary, savefn, title = "Ride", palette = "plasma",
  verticalMultiplier = NA, ppm = NA, elevationShape = 46,
  hrDistance = FALSE, cadDistance = FALSE, hrTime = TRUE,
  cadTime = TRUE, showTime = TRUE, showSummary = TRUE, showStops = TRUE,
  hrLow = 100, hrHigh = 170, hrColorLow = 9, hrColorHigh = 31,
  cadLow = 65, cadTarget = 88, cadCont = TRUE, cadContLow = 60,
  cadContHigh = 100, cadColorLow = 3, cadColorMid = 9,
  cadColorHigh = 15, hrSmoothBW = 6, hrSmoothNN = 5, cadSmoothBW = 20,
  cadSmoothNN = 15, minNumPoints = 1000, imperial = TRUE)
}
\arguments{
\item{track}{data frame or tibble containing track details including
location and elevation data, as well as heartrate and cadence}

\item{summary}{data frame or tibble containing track summary from the
data frames generated by \code{read_ride}}

\item{savefn}{name of output file, format is determined by the file
extension specified, .pdf, .png or .jpg}

\item{title}{title printed over top center of map}

\item{palette}{palette to use in graphic (\code{"plasma"} or \code{"magma"} or
\code{"inferno"} or \code{"viridis"})}

\item{verticalMultiplier}{default vertical exaggeration factor.  Default
ranges from 25 to 60 depending on the length being plotted}

\item{ppm}{override calculated default number of points per mile}

\item{elevationShape}{shape to use for drawing the elevation plot}

\item{hrDistance}{display HR data scaled by distance if available}

\item{cadDistance}{display cadence data scaled by distance if available}

\item{hrTime}{display HR data scaled by time if available}

\item{cadTime}{display cadence data scaled by time if available}

\item{showTime}{draw time axis even if no HR/cad data}

\item{showSummary}{display summary results if supplied}

\item{showStops}{draw row with short and long stops on distance axis}

\item{hrLow}{heartrates below this are same color}

\item{hrHigh}{heartrates above this are same color}

\item{hrColorLow}{set color for hr.low and lower}

\item{hrColorHigh}{set color for hr.highand higher
colors are from same palette as speeds, number is the speed corresponding
to the desired limit on the range of heartrates}

\item{cadLow}{low cadence limit}

\item{cadTarget}{target cadence range minimum}

\item{cadCont}{display cadence as a continuos color map}

\item{cadContLow}{lower cadence limit for continuous color, all
lower cadences are displayed as same color}

\item{cadContHigh}{upper cadence limit for continuous color, all
higher cadences are displayed as same color}

\item{cadColorLow}{set color for cadence at cad.low or below}

\item{cadColorMid}{set color for cadence above low but below target}

\item{cadColorHigh}{set color for cadence above target}

\item{hrSmoothBW}{bandwidth (in seconds) for smoothing kernel
for heartrate data}

\item{hrSmoothNN}{number of points (on each side) to use in smoothing
kernel for heartrate data}

\item{cadSmoothBW}{bandwidth (in seconds) for smoothing kernel
for cadence data}

\item{cadSmoothNN}{number of points (on each side) to use in
smoothing kernel for cadence data}

\item{minNumPoints}{pad out the plot on the right if too short}

\item{imperial}{use mi and ft instead of km and m}
}
\value{
a ggplot object
}
\description{
\code{plot_profile} creates a plot which uses the elevation profile
  and colored bands to compactly display the relationship between terrain,
  speed, cadence and heartrate on a given ride
}
\details{
Generates a plot which compactly displays the elevation profile and its
   relationship with speed, cadence and heartrate.  Also displays clock
   information and shows location and approximate length of stops.
}
