% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_read.R
\name{repairCadence}
\alias{repairCadence}
\title{clean up cadence data for a track}
\usage{
repairCadence(trackdf, cadMax = 160, cadCorrectTooHigh = TRUE,
  cadCorrectNA = FALSE, cadCorrectStopped = TRUE, cadCorrectWindowSec = 7,
  loud = FALSE, ...)
}
\arguments{
\item{trackdf}{data frame or tibble with gps track data}

\item{cadMax}{max credible cadence value, larger values are errors}

\item{cadCorrectTooHigh}{repair excessive cadence values
using triangular-kernel-weighted average of the nearest nonmissing values
in the same segment}

\item{cadCorrectNA}{repair cadence missing values using
triangular-kernel-weighted average of the nearest nonmissing values in
the same segment}

\item{cadCorrectStopped}{repair cadence by setting cadence to 0 when speed
is zero.  This is what would be appropriate if magnet parked near sensor
was generating spurious clicks or if moving pedal while at stoplight}

\item{cadCorrectWindowSec}{window size for kernel smoothing of cadence}

\item{loud}{display actions taken}

\item{...}{parameters for \code{\link{processSegments}},
\code{\link{statsPower}},
\code{\link{statsHeartRate}},
\code{\link{statsGearing}},
\code{\link{statsGrade}},
\code{\link{statsSession}},
\code{\link{statsStops}}}
}
\value{
dataframe with cadence data repaired
}
\description{
\code{repairCadence}  processes a gps track file to correct and
  summarize cadence data
}
\seealso{
\code{\link{read_ride}},
   \code{\link{statsCadence}}
}
