% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_read.R
\name{statsCadence}
\alias{statsCadence}
\title{clean up and summarize cadence statistics for a track}
\usage{
statsCadence(trackdf, cadTrimBegSecs = 15, cadTrimBegMeters = 10,
  cadTrimEndSecs = 20, cadTrimEndMeters = 15, cadMax = 160,
  cadCorrectTooHigh = TRUE, cadCorrectNA = FALSE, cadCorrectWindowSec = 7,
  ...)
}
\arguments{
\item{trackdf}{data frame or tibble with gps track data}

\item{cadTrimBegSecs}{number of seconds after the beginning of a
track segment that are ignored in calculating midsegment avg cadence}

\item{cadTrimBegMeters}{number of meters after the beginning of a
track segment that are ignored in calculating midsegment avg cadence}

\item{cadTrimEndSecs}{number of seconds before the end of a
track segment that are ignored in calculating midsegment avg cadence}

\item{cadTrimEndMeters}{number of meters before the end of a
track segment that are ignored in calculating midsegment avg cadence}

\item{cadMax}{max credible cadence value, larger values are errors}

\item{cadCorrectTooHigh}{repair excessive cadence values
using triangular-kernel-weighted average of the nearest nonmissing values
in the same segment}

\item{cadCorrectNA}{repair cadence missing values using
triangular-kernel-weighted average of the nearest nonmissing values in
the same segment}

\item{cadCorrectWindowSec}{window size for kernel smoothing of cadence}

\item{...}{parameters for \code{\link{processSegments}},
\code{\link{statsPower}},
\code{\link{statsHeartRate}},
\code{\link{statsGearing}},
\code{\link{statsGrade}},
\code{\link{statsSession}},
\code{\link{statsStops}}}
}
\value{
a list containing summary data
}
\description{
\code{statsCadence}  processes a gps track file to correct and
  summarize cadence data
}
\seealso{
\code{\link{read_ride}},\code{\link{processSegments}},
   \code{\link{statsPower}},
   \code{\link{statsHeartRate}},
   \code{\link{statsGearing}},
   \code{\link{statsGrade}},
   \code{\link{statsSession}},
   \code{\link{statsStops}}
}
