% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_read.R
\name{statsGrade}
\alias{statsGrade}
\title{generate grade statistics for a track}
\usage{
statsGrade(trackdf, gradeSmoothWindowMeters = 7, gradeSmoothWindowNN = 10,
  ...)
}
\arguments{
\item{trackdf}{data frame or tibble with gps track data}

\item{gradeSmoothWindowMeters}{bandwidth (meters) for kernel smoothing of
elevation in the grade calculation}

\item{gradeSmoothWindowNN}{number of observations on each side used for
calculating kernel smoothed elevation for the grade calculation}

\item{...}{parameters for \code{\link{processSegments}},
\code{\link{statsCadence}},
\code{\link{statsPower}},
\code{\link{statsHeartRate}},
\code{\link{statsGearing}},
\code{\link{statsSession}},
\code{\link{statsStops}}}
}
\value{
a list containing summary data
}
\description{
\code{statsGrade}  processes a gps track file to summarize grade data
}
\seealso{
\code{\link{read_ride}},\code{\link{processSegments}},
   \code{\link{statsCadence}},
   \code{\link{statsPower}},
   \code{\link{statsHeartRate}},
   \code{\link{statsGearing}},
   \code{\link{statsSession}},
   \code{\link{statsStops}}
}
