% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_read.R
\name{statsPower}
\alias{statsPower}
\title{generate power statistics for a track}
\usage{
statsPower(trackdf, powerCalibrateTime = 0, ...)
}
\arguments{
\item{trackdf}{data frame or tibble with gps track data}

\item{powerCalibrateTime}{number of seconds to ignore in avgpowerPostCal}

\item{...}{parameters for \code{\link{processSegments}},
\code{\link{statsCadence}},
\code{\link{statsHeartRate}},
\code{\link{statsGearing}},
\code{\link{statsGrade}},
\code{\link{statsSession}},
\code{\link{statsStops}}}
}
\value{
a list containing summary data
}
\description{
\code{statsPower}  processes a gps track file to summarize the power data
}
\seealso{
\code{\link{read_ride}},\code{\link{processSegments}},
   \code{\link{statsCadence}},
   \code{\link{statsHeartRate}},
   \code{\link{statsGearing}},
   \code{\link{statsGrade}},
   \code{\link{statsSession}},
   \code{\link{statsStops}}
}
