% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_tracklibs.R
\name{update_gps_variables}
\alias{update_gps_variables}
\title{add/replace ride data in all tbl pairs}
\usage{
update_gps_variables(outdir, fitrootdir, gpxrootdir,
  merge.files = list(c(NA, NA)), fitexcludes = c("bad", "short"),
  gpxexcludes = c("bad", "short", "nosegs"), prefer.gpx = c(""),
  rebuild.all.fit = FALSE, rebuild.all.gpx = FALSE,
  drawprofile = TRUE, drawprofile.both = FALSE, elevationChar = "|",
  plotly = FALSE, rgl = FALSE, rasterDir = NULL,
  localElevFile = "", maptype = "bing", featureDataSource = "Raster",
  townLevel = 9, roadLevel = 3, waterALevel = 4, waterLLevel = 5,
  rglColorScheme = "default", useImageRaster = FALSE,
  plot3DVertScale = 1, cadCont = TRUE, drawmap = TRUE,
  drawmap.both = FALSE, cores = 4, loud = FALSE, ...)
}
\arguments{
\item{outdir}{string containing location of output data pair files}

\item{fitrootdir}{string containing location of root directory for gps .fit files}

\item{gpxrootdir}{string containing location of root directory for gps .gpx files}

\item{merge.files}{list of vectors (order matters within vectors) of .fit
and .gpx files at the top level to be joined}

\item{fitexcludes}{list of substrings within a file name which
will cause the file to be ignored.  occasionally .fit files will be unreadable
by these routines and excluding them will be necessary}

\item{gpxexcludes}{list of substrings within a file
name which will cause the file to be ignored.  occasionally .fit files will
be unreadable by the routines in this package}

\item{prefer.gpx}{vector of strings identifying times to use a .gpx file
rather than the corresponding .fit file}

\item{rebuild.all.fit}{if TRUE, will read all files in subdirectories
of \eqn{fitrootdir}, but not in the directory itself}

\item{rebuild.all.gpx}{if TRUE, will read all files in subdirectories
of \eqn{gpxrootdir}, but not in the directory itself}

\item{drawprofile}{create a .png file in the output directory profiling
the ride}

\item{drawprofile.both}{create a .png file in the output directory profiling
the ride for both the .fit files and .gpx files}

\item{elevationChar}{character to use when printing the elevation profile}

\item{plotly}{if TRUE use plotly to draw 3D track in viewer}

\item{rgl}{if TRUE use rgl to draw 3D track in viewer}

\item{rasterDir}{character location of base directory to load and save raster files}

\item{localElevFile}{file containing raster object with elevations on lat/lon}

\item{maptype}{map to use as background (\code{"maptoolkit-topo"} or
\code{"bing"} or \code{"osm"} or \code{"stamen-terrain"} or
\code{"esri-topo"}  or \code{"stamen-watercolor"} or \code{"mapbox"} or
\code{"esri"} or \code{"osm-public-transport"} or \code{"opencyclemap"}
or \code{"apple-iphoto"} or \code{"skobbler"})}

\item{featureDataSource}{character,  "Raster" to load saved raster data
from directory specified , "none" to show none}

\item{townLevel}{numeric, display towns ranked this number or higher:
3=all towns     5=larger towns (in US >50k)}

\item{roadLevel}{numeric, display roads ranked this number or higher:
2=Service Drive, Bike Path, etc      3=Local Street
4=Secondary Hwy                      5=Primary Hwy/Transit}

\item{waterALevel}{numeric, display areal water ranked this number or higher:
2=res/treatmentpond/pit/quarry       3=lake/pond/swamp/stream
4=class 2 or 3 bigger than 1k ha     5=named lake/pond/swamp/stream
6=large lake/pond/swamp/stream       7=Ocean/Bay/Est/Sound
8=glacier}

\item{waterLLevel}{numeric, display linear water ranked this number or higher:
2=canal/ditch                        3=braided stream
4=stream/river                       5=named stream/river
6=named stream/river containing the string "RIV"}

\item{rglColorScheme}{name of color scheme from
c("default","beach","viridis","plasma","terrain","oleron","snow","oslo",
  "desert","lajolla","niccoli","bright",
  "bing","maptoolkit-topo","nps","apple-iphoto")}

\item{useImageRaster}{logical, use the image raster from saved openStreetmap
colrings of the map surface}

\item{plot3DVertScale}{number which will multiply vertical scaling for plotly/rgl}

\item{cadCont}{display cadence as continuous rather than categorical}

\item{drawmap}{create a .tiff file in the output directory mapping
the ride with speed-varying color}

\item{drawmap.both}{create a .tiff file in the output directory mapping
the ride with speed-varying color for both the .fit files and the .gox files}

\item{cores}{numer of cores (default is #CPUs - 1)}

\item{loud}{display summary of re/segmenting}

\item{...}{parameters passed for track cleaning}
}
\value{
a tbl contining the combined summaries and tracks, with preference
  given to data from .fit files over the corresponding .gpx file, with specified
  exceptions
}
\description{
\code{update_gps_variables} Add/replace ride data in .fit, .gpx and combined
 tbl pairs (summaries,tracks)
}
\details{
Update existing (summaries,tracks) pairs for .fit and .gpx files structured
  with new files at the top level, and files already included in the
  pairs of tbls summarizing the rides are in subdirectories below them.
  there are 3 file pairs created:
    (fitsummary,fittracks), (gpxsummary,gpxtracks) and (gpssummary,gpstracks)
    the last pair contains data from both joined, with preference given to
    .fit-sourced data
}
\seealso{
\code{\link{read_ride}}
}
