% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_visuals.R
\name{plot_elev_profile_plus}
\alias{plot_elev_profile_plus}
\title{plot heartrate, cadence, speed and more using color-driven
  data summaries to convey performance patterns}
\usage{
plot_elev_profile_plus(track, summary, savefn, title = "Ride",
  palette = "plasma", vertical.multiplier = NA, ppm = NA,
  show.stops = TRUE, show.time = TRUE, show.summary = TRUE,
  cad.cont = TRUE, hr.low = 100, hr.high = 170, hr.color.low = 9,
  hr.color.high = 31, cad.low = 65, cad.target = 88, cad.cont.low = 60,
  cad.cont.high = 100, cad.color.low = 3, cad.color.mid = 9,
  cad.color.high = 15, hr.smooth.bw.meters = 6, hr.smooth.nn = 5,
  cad.smooth.bw.meters = 20, cad.smooth.nn = 15, min.numpoints = 1000)
}
\arguments{
\item{track}{data frame or tibble containing track details including
location and elevation data, as well as heartrate and cadence}

\item{summary}{data frame or tibble containing track summary from the
data frames generated by \code{read_ride}}

\item{savefn}{name of output file, format is determined by the file
extension specified, .pdf, .png or .jpg}

\item{title}{title printed over top center of map}

\item{palette}{palette to use in graphic (\code{"plasma"} or \code{"magma"} or
\code{"inferno"} or \code{"viridis"})}

\item{vertical.multiplier}{default vertical exaggeration factor.  Default
ranges from 25 to 60 depending on the length being plotted}

\item{ppm}{override calculated default number of points per mile}

\item{show.stops}{draw row with short and long stops}

\item{show.time}{draw row with 15 minute and hour marks}

\item{show.summary}{display summary results if supplied}

\item{cad.cont}{display cadence as a continuos color map}

\item{hr.low}{heartrates below this are same color}

\item{hr.high}{heartrates above this are same color}

\item{hr.color.low}{set color for hr.low and lower}

\item{hr.color.high}{set color for hr.highand higher
colors are from same palette as speeds, number is the speed corresponding
to the desired limit on the range of heartrates}

\item{cad.low}{low cadence limit}

\item{cad.target}{target cadence range minimum}

\item{cad.cont.low}{lower cadence limit for continuous color, all
lower cadences are displayed as same color}

\item{cad.cont.high}{upper cadence limit for continuous color, all
higher cadences are displayed as same color}

\item{cad.color.low}{set color for cadence at cad.low or below}

\item{cad.color.mid}{set color for cadence above low but below target}

\item{cad.color.high}{set color for cadence above target}

\item{hr.smooth.bw.meters}{bandwidth (in meters) for smoothing kernel
for heartrate data}

\item{hr.smooth.nn}{number of points (on each side) to use in smoothing
kernel for heartrate data}

\item{cad.smooth.bw.meters}{bandwidth (in meters) for smoothing kernel
for cadence data}

\item{cad.smooth.nn}{number of points (on each side) to use in
smoothing kernel for cadence data}

\item{min.numpoints}{pad out the plot on the right if too short}
}
\value{
a ggplot object
}
\description{
\code{plot_elev_profile_plus} creates a plot which uses the elevation profile
  and colored bands to compactly display the relationship between terrain,
  speed, cadence and heartrate on a given ride
}
\details{
Generates a plot which compactly displays the elevation profile and its
   relationship with speed, cadence and heartrate.  Also displays clock
   information and shows location and approximate length of stops.
}
