% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_read.R
\name{read_ride}
\alias{read_ride}
\title{read and clean a gps data file}
\usage{
read_ride(ridefile, tz = "America/Los_Angeles", fixDistance = FALSE,
  loud = FALSE, usefitdc = FALSE, ...)
}
\arguments{
\item{ridefile}{a filenames (.fit or .gpx extension) to process}

\item{tz}{string contaiing timezone for converting from UTC}

\item{fixDistance}{repair nonmonotonicities in distance which are on
segment breaks - this occurs when power is lost or on some device lockups}

\item{loud}{print information about re/segmenting track data}

\item{usefitdc}{use package fitdc to read fit files instead of fitparse}

\item{...}{parameters for \code{\link{processSegments}},
\code{\link{statsCadence}},
\code{\link{statsPower}},
\code{\link{statsHeartRate}},
\code{\link{statsGearing}},
\code{\link{statsGrade}},
\code{\link{statsSession}},
\code{\link{statsStops}}}
}
\value{
a list of three data tibbles:  \eqn{summary} and \eqn{trackpoints}
   and \eqn{session}
   These are linked by integer fields \eqn{startbutton.date}
   and \eqn{startbutton.time}
}
\description{
\code{read_ride}  processes a gps .fit and/or .gpx track file to create
  summary and detail data tibbles for modeling or graphical summary.  Track
  segments are adjusted to remove many false autostop-start and startline
  delay sequences.  Several measures of cadence are calculated (including
  zeros, excluding zeros, and midsegment),spurious and missing cadence values
  are repaired, and summary stats are calculated for analysis.
}
\seealso{
\code{\link{read_ridefiles}},
   \code{\link{processSegments}},
   \code{\link{statsCadence}},
   \code{\link{statsPower}},
   \code{\link{statsHeartRate}},
   \code{\link{statsGearing}},
   \code{\link{statsGrade}},
   \code{\link{statsSession}},
   \code{\link{statsStops}}
}
