% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_tracklibs.R
\name{update_gps_variables}
\alias{update_gps_variables}
\title{add/replace ride data in all tbl pairs}
\usage{
update_gps_variables(outdir, fitrootdir, gpxrootdir, merge.files = list(c(NA,
  NA)), fitexcludes = c("bad", "short"), gpxexcludes = c("bad", "short",
  "nosegs"), prefer.gpx = c(""), rebuild.all.fit = FALSE,
  rebuild.all.gpx = FALSE, drawprofile = TRUE, drawprofile.both = FALSE,
  elevation.char = "|", drawmap = TRUE, drawmap.both = FALSE, cores = 4,
  ...)
}
\arguments{
\item{outdir}{string containing location of output data pair files}

\item{fitrootdir}{string containing location of root directory for gps .fit files}

\item{gpxrootdir}{string containing location of root directory for gps .gpx files}

\item{merge.files}{list of vectors (order matters within vectors) of .fit
and .gpx files at the top level to be joined}

\item{fitexcludes}{list of substrings within a file name which
will cause the file to be ignored.  occasionally .fit files will be unreadable
by these routines and excluding them will be necessary}

\item{gpxexcludes}{list of substrings within a file
name which will cause the file to be ignored.  occasionally .fit files will
be unreadable by the routines in this package}

\item{prefer.gpx}{vector of strings identifying times to use a .gpx file
rather than the corresponding .fit file}

\item{rebuild.all.fit}{if TRUE, will read all files in subdirectories
of \eqn{fitrootdir}, but not in the directory itself}

\item{rebuild.all.gpx}{if TRUE, will read all files in subdirectories
of \eqn{gpxrootdir}, but not in the directory itself}

\item{drawprofile}{create a .png file in the output directory profiling
the ride}

\item{drawprofile.both}{create a .png file in the output directory profiling
the ride for both the .fit files and .gpx files}

\item{elevation.char}{character to use when printing the elevation profile}

\item{drawmap}{create a .tiff file in the output directory mapping
the ride with speed-varying color}

\item{drawmap.both}{create a .tiff file in the output directory mapping
the ride with speed-varying color for both the .fit files and the .gox files}

\item{cores}{numer of cores (default is #CPUs - 1)}

\item{...}{parameters passed for track cleaning}
}
\value{
a tbl contining the combined summaries and tracks, with preference
  given to data from .fit files over the corresponding .gpx file, with specified
  exceptions
}
\description{
\code{update_gps_variables} Add/replace ride data in .fit, .gpx and combined
 tbl pairs (summaries,tracks)
}
\details{
Update existing (summaries,tracks) pairs for .fit and .gpx files structured
  with new files at the top level, and files already included in the
  pairs of tbls summarizing the rides are in subdirectories below them.
  there are 3 file pairs created:
    (fitsummary,fittracks), (gpxsummary,gpxtracks) and (gpssummary,gpstracks)
    the last pair contains data from both joined, with preference given to
    .fit-sourced data
}
\seealso{
\code{\link{read_ride}}
}
